pyproject.toml
[project]
name = "mccag"
version = "0.1.2"
description = "Minecraft Cute Avatar Generator"
authors = [
    {name = "SerinaNya", email = "34389622+SerinaNya@users.noreply.github.com"},
]
dependencies = [
    "httpx[http2,socks]>=0.27.0",
    "pydantic>=2.8.2",
    "pydantic-settings>=2.3.4",
    "yggdrasil-mc>=0.2.4",
    "fastapi>=0.111.1",
    "uvicorn[standard]>=0.30.3",
    "pillow>=10.4.0",
    "loguru>=0.7.2",
    "richuru>=0.1.1",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "GPL-3.0-or-later"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm.scripts]
dev = { cmd = "uvicorn mccag.main:app --port 8023 --reload", help = "Start the server" }

[tool.pdm]
distribution = true

[tool.ruff]
extend-exclude = []
line-length = 120
target-version = "py312"